+++ com.unity.render-pipelines.high-definition/Runtime/Material/Decal/DecalSystem.cs
@@ -165,19 +165,27 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
         private List<TextureScaleBias> m_TextureList = new List<TextureScaleBias>();
 
+
         private class DecalSet
         {
             public void InitializeMaterialValues()
             {
                 if (m_Material == null)
                     return;
-                m_Diffuse.Initialize(m_Material.GetTexture("_BaseColorMap"), Vector4.zero); 
-                m_Normal.Initialize(m_Material.GetTexture("_NormalMap"), Vector4.zero); 
-                m_Mask.Initialize(m_Material.GetTexture("_MaskMap"), Vector4.zero);
-                m_Blend = m_Material.GetFloat("_DecalBlend");
-                m_AlbedoContribution = m_Material.GetFloat("_AlbedoMode");
-                m_BaseColor = m_Material.GetVector("_BaseColor");
-                m_BlendParams = new Vector3(m_Material.GetFloat("_NormalBlendSrc"), m_Material.GetFloat("_MaskBlendSrc"), m_Material.GetFloat("_MaskBlendMode"));                
+                m_IsHDRenderPipelineDecal = (m_Material.shader.name == "HDRP/Decal");
+
+                if (m_IsHDRenderPipelineDecal)
+                {
+                    m_Diffuse.Initialize(m_Material.GetTexture("_BaseColorMap"), Vector4.zero);
+                    m_Normal.Initialize(m_Material.GetTexture("_NormalMap"), Vector4.zero);
+                    m_Mask.Initialize(m_Material.GetTexture("_MaskMap"), Vector4.zero);
+                    m_Blend = m_Material.GetFloat("_DecalBlend");
+                    m_AlbedoContribution = m_Material.GetFloat("_AlbedoMode");
+                    m_BaseColor = m_Material.GetVector("_BaseColor");
+                    m_BlendParams = new Vector3(m_Material.GetFloat("_NormalBlendSrc"), m_Material.GetFloat("_MaskBlendSrc"), m_Material.GetFloat("_MaskBlendMode"));
+                    m_RemappingAOS = new Vector4(m_Material.GetFloat("_AORemapMin"), m_Material.GetFloat("_AORemapMax"), m_Material.GetFloat("_SmoothnessRemapMin"), m_Material.GetFloat("_SmoothnessRemapMax"));
+                    m_ScalingMAB = new Vector4(m_Material.GetFloat("_MetallicScale"), 0.0f, m_Material.GetFloat("_DecalMaskMapBlueScale"), 0.0f);
+                }
             }
 
             public DecalSet(Material material)
@@ -410,7 +418,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                         normalToWorldBatch[instanceCount].SetRow(3, m_CachedUVScaleBias[decalIndex]);
 
                         // clustered forward data
-                        if(m_CachedAffectsTransparency[decalIndex])
+                        if (m_CachedAffectsTransparency[decalIndex])
                         { 
                             m_DecalDatas[m_DecalDatasCount].worldToDecal = decalToWorldBatch[instanceCount].inverse;
                             m_DecalDatas[m_DecalDatasCount].normalToWorld = normalToWorldBatch[instanceCount];
@@ -442,7 +450,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                 }
 
                 // only add if any projectors in this decal set affect transparency, doesn't actually allocate textures in the atlas yet, this is because we want all the textures in the list so we can optimize the packing
-                if( anyAffectTransparency)
+                if (anyAffectTransparency)
                 { 
                     AddToTextureList(ref instance.m_TextureList);
                 }
@@ -489,7 +497,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                 int totalToDraw = m_NumResults;
                 HDRenderPipelineAsset hdrp = GraphicsSettings.renderPipelineAsset as HDRenderPipelineAsset;
                 bool perChannelMask = hdrp.renderPipelineSettings.decalSettings.perChannelMask;
-                int shaderPass = perChannelMask ? (int)m_Material.GetFloat("_MaskBlendMode") : (int)Decal.MaskBlendFlags.Smoothness; // relies on the order shader passes are declared in decal.shader and decalUI.cs
+                int shaderPass = perChannelMask ? MaskBlendMode : (int)Decal.MaskBlendFlags.Smoothness; // relies on the order shader passes are declared in decal.shader and decalUI.cs
               
                 for (; batchIndex < m_NumResults / kDrawIndexedBatchSize; batchIndex++)
                 {
@@ -507,7 +515,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
 
             public void Cleanup()
             {
-                if(m_CullingGroup != null)
+                if (m_CullingGroup != null)
                 {
                     CullingGroupManager.instance.Free(m_CullingGroup);
                 }
@@ -537,6 +545,24 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
                 }
             }
 
+            public int MaskBlendMode
+            {
+                get
+                {
+                    if (m_IsHDRenderPipelineDecal)
+                    {
+                        return (int)this.m_Material.GetFloat("_MaskBlendMode");
+                    }
+                    else
+                    {
+                        // TODO
+                        return 0;
+                    }
+                }
+            }
+
+
+
             private List<Matrix4x4[]> m_DecalToWorld = new List<Matrix4x4[]>();
             private List<Matrix4x4[]> m_NormalToWorld = new List<Matrix4x4[]>();
 
@@ -556,7 +582,9 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             private float m_AlbedoContribution = 0;
             private Vector4 m_BaseColor;
             private Vector3 m_BlendParams;
-            
+
+            private bool m_IsHDRenderPipelineDecal;
+
             TextureScaleBias m_Diffuse = new TextureScaleBias();
             TextureScaleBias m_Normal = new TextureScaleBias();
             TextureScaleBias m_Mask = new TextureScaleBias();